some fix

+++ src/main/java/loglang/LoglangFactory.java
@@ -95,9 +95,11 @@ public class LoglangFactory {
             String pathName = path.toString();
 
             // delete pattern file before shutdown
-            Runtime.getRuntime().addShutdownHook(
-                    new Thread(() -> new File(pathName).delete())
-            );
+            if(Utils.checkProperty("deletePattern", true)) {
+                Runtime.getRuntime().addShutdownHook(
+                        new Thread(() -> new File(pathName).delete())
+                );
+            }
 
             return GrammarFile.loadGrammarFile(pathName, NezOption.newDefaultOption()).newGrammar("File");
         } catch(IOException e) {
